# :zap: Feat: Optimize state manager peformance and caching