--- a/.changeset/fair-weeks-glow.md
+++ b/.changeset/fair-weeks-glow.md
@@ -1,0 +1,19 @@
+---
+"@tevm/memory-client": patch
+"@tevm/base-client": patch
+"@tevm/state": patch
+---
+
+Made major improvements to tevm performance
+
+before: 2.38s
+after: 0.83s
+
+before: 6.69s
+after: 0.683s
+
+before: 8.42s
+after: 0.219s
+
+before: 4.07s
+after: 0.01s