--- a/docs/src/content/docs/reference/@tevm/state/functions/dumpStorage.md
+++ b/docs/src/content/docs/reference/@tevm/state/functions/dumpStorage.md
@@ -5,7 +5,7 @@ prev: false
 title: "dumpStorage"
 ---
 
-> **dumpStorage**(`baseState`): (`address`) => `Promise`\<[`StorageDump`](/reference/tevm/common/interfaces/storagedump/)\>
+> **dumpStorage**(`baseState`, `skipFetchingFromFork`?): (`address`) => `Promise`\<[`StorageDump`](/reference/tevm/common/interfaces/storagedump/)\>
 
 Dumps the RLP-encoded storage values for an `account` specified by `address`.
 Keys are the storage keys, values are the storage values as strings.
@@ -14,6 +14,8 @@ Both are represented as `0x` prefixed hex strings.
 ## Parameters
 
 • **baseState**: [`BaseState`](/reference/tevm/state/type-aliases/basestate/)
+
+• **skipFetchingFromFork?**: `boolean`
 
 ## Returns