--- a/packages/base-client/src/createBaseClient.js
+++ b/packages/base-client/src/createBaseClient.js
@@ -78,7 +78,6 @@ export const createBaseClient = (options = {}) => {
 		return /** @type {any}*/ (client)
 	}
 
-	console.time('chainIdPromise-ready')
 	const chainIdPromise = (async () => {
 		if (options?.common) {
 			return options?.common.id
@@ -91,11 +90,9 @@ export const createBaseClient = (options = {}) => {
 		return DEFAULT_CHAIN_ID
 	})().then((chainId) => {
 		logger.debug({ chainId }, 'Creating client with chainId')
-		console.timeEnd('chainIdPromise-ready')
 		return BigInt(chainId)
 	})
 
-	console.time('blockTagPromise-ready')
 	const blockTagPromise = (async () => {
 		if (options.fork === undefined) {
 			// this is ultimately unused
@@ -109,13 +106,8 @@ export const createBaseClient = (options = {}) => {
 			return latestBlockNumber
 		}
 		return options.fork.blockTag
-	})().then((blockTag) => {
-		console.timeEnd('blockTagPromise-ready')
-		console.time('blockchain-ready')
-		return blockTag
-	})
+	})()
 
-	console.time('chainCommonPromise-ready')
 	const chainCommonPromise = chainIdPromise
 		.then((chainId) => {
 			if (options.common) {
@@ -139,7 +131,6 @@ export const createBaseClient = (options = {}) => {
 		.then((common) => {
 			// ALWAYS Copy common so we don't modify the global instances since it's stateful!
 			const copy = common.copy()
-			console.timeEnd('chainCommonPromise-ready')
 			return copy
 		})
 
@@ -166,7 +157,6 @@ export const createBaseClient = (options = {}) => {
 		})
 		.then((chain) => chain.getCanonicalHeadBlock())
 		.then(async (headBlock) => {
-			console.time('stateManager-genesis')
 			const stateRootHex = bytesToHex(headBlock.header.stateRoot)
 			const restoredState = options.persister?.restoreState()
 			if (restoredState) {
@@ -215,9 +205,7 @@ export const createBaseClient = (options = {}) => {
 					]),
 				),
 			}
-			console.timeEnd('stateManager-genesis')
 			const opts = await getStateManagerOpts()
-			console.time('stateManager-ready')
 			return createStateManager({
 				...opts,
 				currentStateRoot: stateRootHex,
@@ -258,30 +246,11 @@ export const createBaseClient = (options = {}) => {
 	})
 
 	const readyPromise = (async () => {
-		console.log('calling ready...')
 		await blockchainPromise.then((b) => b.ready())
-		console.log('blockchain...')
 		await stateManagerPromise.then((b) => b.ready())
-		console.log('stateManager...')
 		await vmPromise.then((vm) => vm.ready())
-		console.log('vm...')
 		return /** @type {true}*/ (true)
 	})()
-	blockchainPromise.then(() => {
-		console.timeEnd('blockchain-ready')
-	})
-	stateManagerPromise.then(() => {
-		console.timeEnd('stateManager-ready')
-		console.time('vm-ready')
-	})
-	vmPromise.then(() => {
-		console.timeEnd('vm-ready')
-		console.time('vm-ready-lag')
-	})
-	readyPromise.then(() => {
-		console.timeEnd('vm-ready-lag')
-		console.timeEnd('baseClient-ready')
-	})
 
 	/**
 	 * Create and return the baseClient