--- a/packages/state/src/actions/dumpCannonicalGenesis.js
+++ b/packages/state/src/actions/dumpCannonicalGenesis.js
@@ -29,15 +29,16 @@ export const dumpCanonicalGenesis = (baseState) => async () => {
 
 	for (const address of accountAddresses) {
 		const hexAddress = getAddress(`0x${address}`)
-		const account = await getAccount(baseState, true)(EthjsAddress.fromString(hexAddress))
+		const ethAddress = EthjsAddress.fromString(hexAddress)
+		const account = await getAccount(baseState, true)(ethAddress)
 
 		if (account === undefined) {
 			baseState.logger.debug({ address: hexAddress }, 'Warning: Account in accountAddresses not found')
 		}
 		if (account !== undefined) {
-			const storage = await dumpStorage(baseState, true)(EthjsAddress.fromString(hexAddress))
+			const storage = await dumpStorage(baseState, true)(ethAddress)
 
-			const deployedBytecode = await getContractCode(baseState, true)(EthjsAddress.fromString(hexAddress))
+			const deployedBytecode = await getContractCode(baseState, true)(ethAddress)
 
 			const dump = {
 				nonce: account.nonce,
@@ -45,7 +46,7 @@ export const dumpCanonicalGenesis = (baseState) => async () => {
 				storageRoot: bytesToHex(account.storageRoot),
 				codeHash: bytesToHex(account.codeHash),
 				storage,
-				deployedBytecode: toHex(deployedBytecode),
+				...(baseState.caches.contracts.has(ethAddress) ? { deployedBytecode: toHex(deployedBytecode) } : {}),
 			}
 
 			baseState.logger.debug({ address: hexAddress, ...dump }, 'dumping address')