--- a/packages/state/src/state-types/StateAction.ts
+++ b/packages/state/src/state-types/StateAction.ts
@@ -1,4 +1,7 @@
 import type { BaseState } from '../BaseState.js'
 import type { StateManager } from '../StateManager.js'
 
-export type StateAction<T extends keyof StateManager> = (baseState: BaseState) => StateManager[T]
+export type StateAction<T extends keyof StateManager> = (
+	baseState: BaseState,
+	skipFetchingFromFork?: boolean,
+) => StateManager[T]