… docs/src/content/docs/reference/@tevm/state/functions/dumpStorageRange.md
… docs/src/content/docs/reference/@tevm/state/functions/generateCanonicalGenesis.md
… docs/src/content/docs/reference/@tevm/state/functions/getAccount.md
… docs/src/content/docs/reference/@tevm/state/functions/getAccountAddresses.md
… docs/src/content/docs/reference/@tevm/state/functions/getAppliedKey.md
… docs/src/content/docs/reference/@tevm/state/functions/getContractCode.md
… docs/src/content/docs/reference/@tevm/state/functions/getContractStorage.md
… docs/src/content/docs/reference/@tevm/state/functions/getProof.md
… docs/src/content/docs/reference/@tevm/state/functions/getStateRoot.md
… docs/src/content/docs/reference/@tevm/state/functions/hasStateRoot.md
… docs/src/content/docs/reference/@tevm/state/functions/modifyAccountFields.md
… docs/src/content/docs/reference/@tevm/state/functions/originalStorageCache.md
… docs/src/content/docs/reference/@tevm/state/functions/putAccount.md
… docs/src/content/docs/reference/@tevm/state/functions/putContractCode.md
… docs/src/content/docs/reference/@tevm/state/functions/putContractStorage.md
… docs/src/content/docs/reference/@tevm/state/functions/revert.md
… docs/src/content/docs/reference/@tevm/state/functions/setStateRoot.md
… docs/src/content/docs/reference/@tevm/state/type-aliases/StateAction.md
… packages/memory-client/src/test/__snapshots__/viemPublicActions.spec.ts.snap
… packages/state/docs/classes/ContractCache.md
… packages/state/docs/functions/checkpoint.md
… packages/state/docs/functions/clearCaches.md
… packages/state/docs/functions/clearContractStorage.md
… packages/state/docs/functions/commit.md
… packages/state/docs/functions/deleteAccount.md
… packages/state/docs/functions/dumpCanonicalGenesis.md
… packages/state/docs/functions/dumpStorage.md
… packages/state/docs/functions/dumpStorageRange.md
… packages/state/docs/functions/generateCanonicalGenesis.md
… packages/state/docs/functions/getAccount.md