### TL;DR
Fixes a bug where eth_getCode would get triggered on every account in forked mode everytime we committed state. Now  eth_getCode runs at most one time along with any other empty cache situation.

### What changed?
Initialization of a MemoryClient in forked mode (ready)

before: 2.38s
after: 0.83s

Deploying a contract as a call (deploy-call)

before: 6.69s
after: 0.683s

Mining a tx with this deploy (mine)

before: 8.42s
after: 0.219s

Executing a call against our deployed contract (call)

before: 4.07s
after: 0.01s

---

## Description

_Concise description of proposed changes_

## Testing

Explain the quality checks that have been done on the code changes

## Additional Information

- [ ] I read the [contributing docs](../docs/contributing.md) (if this is your first contribution)

Your ENS/address:



<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **Performance Improvements**
  - Major performance enhancements to reduce execution times for various components.

- **New Features**
  - Added detailed timing logs for asynchronous operations to improve performance monitoring.
  - Introduced new methods and parameters to enhance functionality and control flow.

- **Bug Fixes**
  - Added checks to ensure valid contracts before retrieving storage, preventing errors.

- **Refactor**
  - Improved handling of asynchronous operations and readiness checks.
  - Introduced structured approaches for handling and caching data.

- **Tests**
  - Added new test cases and updated existing ones to cover new functionalities and bug fixes.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->